Checked
git-svn-id: https://svn.sourceforge.jp/svnroot/nucleus-jp/plugin@1057 1ca29b6e-896d-4ea0-84a5-967f57386b96

+++ NP_gallery/trunk/gallery/forms.php
@@ -293,6 +293,7 @@ function addPictureForm($albumid = 0, $num_files = 0) {
 
 function addpictureformjupload($albumid = 0, $num_files = 0) {
 	global $NPG_CONF,$CONF;
+	if (!preg_match('/^([a-z0-9_]+|`[^`]+`)$/i',$NPG_CONF['temp_table'])) exit;
 	$exist_temp_table = mysql_query('SELECT 1 FROM '.$NPG_CONF['temp_table'].' LIMIT 0');
 	if ($exist_temp_table) sql_query('drop table '. $NPG_CONF['temp_table']);
 
@@ -358,6 +359,7 @@ function addTempPictureForm($albumid = 0) {
 	global $NPG_CONF, $gmember,$manager,$CONF,$NP_BASE_DIR;
 	
 	$NPG_CONF = getNPGConfig();
+	if (!preg_match('/^([a-z0-9_]+|`[^`]+`)$/i',$NPG_CONF['temp_table'])) exit;
 	$table_name = $NPG_CONF['temp_table'];
 	
 	$promo_allowed = false;
